publisher double serialize, need to deserialize before publuish in transaction service, and edit publisher parameters. no routing key incoming needed

+++ AccountService/src/AccountService.Application/DomainEventHandlers/TransferFailedDomainEventHandler.cs
@@ -24,7 +24,7 @@ public class TransferFailedDomainEventHandler : INotificationHandler<DomainEvent
             domainEvent.FromAccountNumber,
             domainEvent.ToAccountNumber,
             domainEvent.Amount,
-            domainEvent.Currency
+            (int)domainEvent.Currency
         );
 
         await _outboxWriter.EnqueueAsync(integrationEvent, ct);
